requirements.txt
fastapi
setuptools
sqlalchemy
asyncpg
aioredis
python-dotenv
pyjwt
passlib